pyproject.toml
[tool.poetry]
name = "edm_su_api"
version = "1.5.1"
description = "edm.su backend"
authors = ["Egor <eagluhih@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.9"
fastapi = { version = "0.94.1", extras = ["all"] }
Pillow = "^9.4.0"
SQLAlchemy = { extras = ["asyncio"], version = "^2.0" }
PyJWT = "^2.6.0"
gunicorn = "^20.1.0"
python-slugify = { extras = ["unidecode"], version = "^8.0" }
aiohttp = "^3.8.1"
psycopg2 = "^2.9.5"
aioboto3 = "^10.3.0"
meilisearch-python-async = "1.1.0"
typing-extensions = "^4.5.0"
greenlet = "^1.1.3.post0"
asyncpg = "^0.27.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-asyncio = "^0"
Faker = "^15.3.4"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest-dotenv = "^0.5.2"
pytest-httpx = "^0.22.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
types-python-slugify = "^8.0.0"
mypy = "^1.1.1"
ruff = "^0.0.262"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
exclude = ["alembic/*"]
plugins = ["pydantic.mypy"]

[tool.ruff]
select = [
  "E",      # pycodestyle Error
  "F",      # Pyflakes
  "W",      # pycodestyle Warnings
  "C90",    # mccabe
  "I",      # isort
  "N",      #pep8-naming
  "UP",     # pyupgrade
  "YTT",    # flake8-2020
  "ANN",    # flake8-annotations
  "S",      # flake8-bandit
  "BLE",    # flake8-blind-except
  "FBT",    # flake8-boolean-trap
  "B",      # flake8-bugbear
  "A",      # flake8-builtins
  "COM",    # flake8-commas
  "C4",     # flake8-comprehensions
  "DTZ001", # flake8-datetimez
  "T10",    # flake8-debugger
  "EM",     # flake8-errmsg
  "ISC",    # flake8-implicit-str-concat
  "ICN",    # flake8-import-conventions
  "G",      # flake8-logging-format
  "INP",    # flake8-no-pep420
  "PIE",    # flake8-pie
  "T20",    # flake8-print,
  "PT",     # flake8-pytest-style
  "Q",      # flake8-quotes
  "RSE",    # flake8-raise
  "RET",    # flake8-return
  "SLF",    # flake8-self
  "SIM",    # flake8-simplify
  "TID",    # flake8-tidy-imports
  "TCH",    # flake8-type-checking
  "ARG",    # flake8-unused-arguments
  "PTH",    # flake8-use-pathlib
  "ERA",    # eradicate
  "PGH",    # pygrep-hooks
  "PL",     # Pylint
  "PLE",    # Error
  "PLR",    # Refactor
  "PLW",    # Warninig
  "TRY",    # tryceratops
  "RUF",    # Ruff-specific rules
]
ignore = ["A003", "RUF001", "RUF002", "RUF003", "TRY301"]

line-length = 79

target-version = "py310"

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "INP001", "ARG", "PLR0913", "S106"]
"alembic/*" = ["ANN", "INP001"]

[tool.ruff.flake8-bugbear]
extend-immutable-calls = [
  "fastapi.Depends",
  "fastapi.Query",
  "fastapi.Header",
  "fastapi.Body",
  "fastapi.Path",
  "fastapi.File",
  "datetime.date.today",
]

[tool.black]
line-length = 79
target-version = ["py310"]
